refactor(runtime): ThemeManager 移入 runtime，消除 runtime → ui 反向依赖
